refactor: remove deprecated agent components and streamline routing
- Deleted obsolete agent files and types to enhance code clarity and maintainability.
- Updated routing in `index.ts` to replace the removed agent routes with the new `agentV2Routes`.
- Consolidated agent-related imports and adjusted the application structure for improved organization.
- Enhanced documentation to reflect the latest changes in agent functionality and routing strategies.
