+++ api/src/index.ts
@@ -11,7 +11,6 @@ import { databaseRoutes } from "./routes/database";
 import { dataSourceRoutes } from "./routes/sources";
 import { customPromptRoutes } from "./routes/custom-prompt";
 import { chatsRoutes } from "./routes/chats";
-import { agentRoutes } from "./routes/agent";
 import { agentV2Routes } from "./routes/agent-v2.routes";
 import { authRoutes } from "./auth/auth.controller";
 import { connectDatabase } from "./database/schema";
@@ -97,8 +96,7 @@ app.route("/api/workspaces/:workspaceId/flows", flowRoutes);
 app.route("/api/run", executeRoutes);
 app.route("/api/execute", executeRoutes);
 app.route("/api/database", databaseRoutes);
-app.route("/api/agent", agentRoutes);
-app.route("/api/agent/v2", agentV2Routes);
+app.route("/api/agent", agentV2Routes);
 app.route("/api/connectors", connectorRoutes);
 app.route("/api/databases", databaseSchemaRoutes);
 app.route("/api/workspaces/:workspaceId/databases", databaseTreeRoutes);
