+++ app/src/store/settingsStore.ts
@@ -1,18 +1,12 @@
 /**
  * Settings Store
- * Manages application settings including AI agent version preference and model selection
+ * Manages application settings including AI model selection
  */
 
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
 
-export type AgentVersion = "v1" | "v2";
-
 interface SettingsState {
-  // Agent version preference
-  agentVersion: AgentVersion;
-  setAgentVersion: (version: AgentVersion) => void;
-
   // AI Model selection
   selectedModelId: string;
   setSelectedModelId: (modelId: string) => void;
@@ -25,12 +19,8 @@ interface SettingsState {
 export const useSettingsStore = create<SettingsState>()(
   persist(
     set => ({
-      // Default to v1 (stable) - users can opt-in to v2 (beta)
-      agentVersion: "v1",
-      setAgentVersion: version => set({ agentVersion: version }),
-
-      // Default to GPT-5.2 (latest flagship model)
-      selectedModelId: "gpt-5.2",
+      // Default model
+      selectedModelId: "gpt-4o",
       setSelectedModelId: modelId => set({ selectedModelId: modelId }),
 
       // Theme defaults to system
@@ -41,7 +31,6 @@ export const useSettingsStore = create<SettingsState>()(
       name: "settings-storage",
       // Only persist specific fields
       partialize: state => ({
-        agentVersion: state.agentVersion,
         selectedModelId: state.selectedModelId,
         theme: state.theme,
       }),
